requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
Pillow>=9.0.0
requests>=2.28.0 